refactor/#6 [remove] : 온보딩 음주여부, 종교 기능 제거
