--- app/src/main/java/com/devnuts/ruflu/ui/onboarding/fragment/OnboardingFancyFragment.kt
@@ -1,51 +1,0 @@
-package com.devnuts.ruflu.ui.onboarding.fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
-import com.devnuts.ruflu.R
-import com.devnuts.ruflu.databinding.FragmentOnboardingFancyBinding
-import com.devnuts.ruflu.ui.onboarding.viewmodel.FormViewModel
-
-/* 온보딩 - 7 : 스타일 */
-class OnboardingFancyFragment : Fragment() {
-    private var _binding: FragmentOnboardingFancyBinding? = null
-    private val binding get() = _binding ?: error("View를 참조하기 위해 binding이 초기화 되지 x")
-    private val viewModel: FormViewModel by viewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentOnboardingFancyBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        progressBar()
-        initializeView()
-    }
-
-    private fun progressBar() {
-        val curProgressBar = binding.pbLoading
-        curProgressBar.progress = 58
-    }
-
-    private fun initializeView() {
-        binding.formNextBnt.setOnClickListener {
-            findNavController().navigate(R.id.action_formFragment_to_eduFragment)
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
