+++ app/src/main/java/com/devnuts/ruflu/ui/onboarding/fragment/eight/OnboardingPhotoFragment.kt
@@ -8,38 +8,32 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.devnuts.ruflu.R
-import com.devnuts.ruflu.databinding.FragmentOnboardingAlcoholBinding
-import com.devnuts.ruflu.ui.onboarding.viewmodel.AlcohViewModel
+import com.devnuts.ruflu.databinding.FragmentOnboardingPhotoBinding
 
-/* 온보딩 - 11 : 주량 정도 */
-class OnboardingAlcoholFragment : Fragment() {
-    private var _binding: FragmentOnboardingAlcoholBinding? = null
+/* 온보딩 - 12 : 사진 업로드 */
+class OnboardingPhotoFragment : Fragment() {
+    private var _binding: FragmentOnboardingPhotoBinding? = null
     private val binding get() = _binding!!
-    private val viewModel: AlcohViewModel by viewModels()
+    //private val viewModel: PhotoUpldViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentOnboardingAlcoholBinding.inflate(inflater, container, false)
+        _binding = FragmentOnboardingPhotoBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        progressBar()
+
         initializeView()
     }
 
-    private fun progressBar() {
-        val curProgressBar = binding.pbLoading
-        curProgressBar.progress = 91
-    }
-
     private fun initializeView() {
-        binding.alcohNextBnt.setOnClickListener {
-            findNavController().navigate(R.id.action_alcohFragment_to_photoUpldFragment)
+        binding.photoUpldNextBnt.setOnClickListener {
+            findNavController().navigate(R.id.action_photoFragment_to_loop)
         }
     }
