+++ app/src/main/java/com/devnuts/ruflu/ui/onboarding/fragment/seven/OnboardingFancyFragment.kt
@@ -1,4 +1,4 @@
-package com.devnuts.ruflu.ui.onboarding.fragment
+package com.devnuts.ruflu.ui.onboarding.fragment.six
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,39 +8,32 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.devnuts.ruflu.R
-import com.devnuts.ruflu.databinding.FragmentOnboardingPhotoBinding
-import com.devnuts.ruflu.ui.onboarding.viewmodel.PhotoUpldViewModel
+import com.devnuts.ruflu.databinding.FragmentOnboardingFancyBinding
 
-/* 온보딩 - 12 : 사진 업로드 */
-class OnboardingPhotoFragment : Fragment() {
-    private var _binding: FragmentOnboardingPhotoBinding? = null
+/* 온보딩 - 7 : 스타일 */
+class OnboardingFancyFragment : Fragment() {
+    private var _binding: FragmentOnboardingFancyBinding? = null
     private val binding get() = _binding!!
-    private val viewModel: PhotoUpldViewModel by viewModels()
+    //private val viewModel: FormViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentOnboardingPhotoBinding.inflate(inflater, container, false)
+        _binding = FragmentOnboardingFancyBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        progressBar()
         initializeView()
     }
 
-    private fun progressBar() {
-        val curProgressBar = binding.pbLoading
-        curProgressBar.progress = 100
-    }
-
     private fun initializeView() {
-        binding.photoUpldNextBnt.setOnClickListener {
-            findNavController().navigate(R.id.action_alcohFragment_to_loop)
+        binding.formNextBnt.setOnClickListener {
+            findNavController().navigate(R.id.action_fancyFragment_to_photoFragment)
         }
     }
