--- app/src/main/res/layout/fragment_onboarding_religion.xml
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-        <variable
-            name="viewModel"
-            type="com.devnuts.ruflu.ui.onboarding.fragment.OnboardingReligionFragment" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/ruflu_almostWhite">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/cl_loading"
-            android:layout_width="match_parent"
-            android:layout_height="80dp"
-            app:layout_constraintTop_toTopOf="parent"
-            android:background="@color/ruflu_almostWhite"
-            >
-            <ProgressBar
-                android:id="@+id/pb_loading"
-                android:layout_width="0dp"
-                android:layout_height="10dp"
-                android:indeterminate="false"
-                android:layout_centerInParent="true"
-                android:progress="0"
-                android:max="100"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toStartOf="@id/tv_loading"
-                style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-                android:progressDrawable="@drawable/style_progressbar"
-                android:layout_marginLeft="20dp"
-                />
-
-            <TextView
-                android:id="@+id/tv_loading"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="10/12"
-                android:includeFontPadding="false"
-                style="@style/TextStyle5"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toEndOf="@id/pb_loading"
-                app:layout_constraintEnd_toEndOf="parent"
-                android:layout_marginLeft="10dp"
-                android:layout_marginRight="20dp"
-                />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <TextView
-            android:id="@+id/tv_effect"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="20dp"
-            android:text="종교를 선택해주세요"
-            android:includeFontPadding="false"
-            style="@style/TextStyle2"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/cl_loading" />
-
-        <TextView
-            android:id="@+id/religNextBnt"
-            android:layout_width="match_parent"
-            android:layout_height="54dp"
-            android:layout_marginLeft="20dp"
-            android:layout_marginTop="24dp"
-            android:layout_marginRight="20dp"
-            android:layout_marginBottom="24dp"
-            android:background="@drawable/onboard_button_selector"
-            android:gravity="center"
-            android:text="다음"
-            android:textColor="@color/ruflu_white"
-            app:layout_constraintBottom_toBottomOf="parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
